feat(Doxygens)!: documented the whole lib
	- add complete doxygens in every headers include/ and private/.
	- fix: date format in for every file should now be correct and in : year-month-day.
	- note: errors may have slipped in the documentation not being exactly like the functions workflow.
	- chore: fix the README.md to be nice and clean.

+++ dkz/src/math/my_sign.c
@@ -5,7 +5,7 @@
 ** my_sign
 */
 
-int my_sign(int x)
+int my_sign(double x)
 {
     if (x < 0)
         return -1;
